feat(norm): register engines lazily so Norm works on edge runtimes
Norm.ts value-imported the four engine classes from the drivers root
barrel purely to switch on `database.dialect`, putting the native SQLite
adapter in every consumer's runtime graph. Replace the switch with a
dialect -> factory registry fed by per-dialect side-effect modules, add
the three fetch-only dialects (neon/turso/d1), and split the barrel into
`mod.ts` (registers all seven, unchanged for existing users) and
`core.ts` / `@tundralibs/norm/core` (registers none).
